# src/config/swagger.yaml
swagger: "2.0"
info:
  description: "Controle Simples de Postagens com Usuários"
  version: "1.0.0"
  title: "posts_control"
basePath: "/api/v1"
consumes:
  - "application/json"
produces:
  - "application/json"
...
definitions:
  user:
    additionalProperties: false
    type: "object"
    required:
      - "full_name"
      - "user_email"
      - "user_password"
    properties:
      id:
        type: "integer"
      full_name:
        type: "string"
      user_email:
        type: "string"
      user_password:
        type: "string"
...
paths:
  /user:
    get:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: list User
      operationId: listUserHandler
      parameters:
        - name: user_id
          in: query
          type: string
          required: false
      responses:
        200:
          description: "success"
          schema:
            type: array
            items:
              $ref: "#/definitions/user"
    post:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: create user
      operationId: createUserHandler
      parameters:
        - in: body
          name: user
          description: create user
          required: true
          schema:
            $ref: "#/definitions/user"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/user"
    put:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: update user
      operationId: updateUserHandler
      parameters:
        - in: body
          name: user
          description: update user
          schema:
            additionalProperties: false
            type: "object"
            required:
              - "id"
              - "full_name"
              - "user_email"
              - "user_password"
            properties:
              id:
                type: "integer"
              full_name:
                type: "string"
              user_email:
                type: "string"
              user_password:
                type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/user"
    delete:
      tags:
        - "Users"
      x-swagger-router-controller: app
      summary: delete user
      operationId: deleteUserHandler
      parameters:
        - name: user_id
          in: query
          type: integer
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/user"
...